+++ crates/client/sync/src/l2.rs
@@ -4,6 +4,12 @@ use std::pin::pin;
 use std::sync::{Arc, Mutex};
 use std::time::Instant;
 
+use crate::commitments::compute_state_root;
+use crate::convert::{convert_and_verify_block, convert_and_verify_class};
+use crate::fetch::fetchers::{fetch_block_and_updates, FetchBlockId, L2BlockAndUpdates};
+use crate::fetch::l2_fetch_task;
+use crate::metrics::block_metrics::BlockMetrics;
+use crate::utility::trim_hash;
 use anyhow::{bail, Context};
 use dc_db::db_metrics::DbMetrics;
 use dc_db::DeoxysBackend;
@@ -15,6 +21,9 @@ use dp_class::ConvertedClass;
 use dp_convert::ToStarkFelt;
 use dp_state_update::StateDiff;
 use dp_transactions::TransactionTypeError;
+use dp_utils::{
+    channel_wait_or_graceful_shutdown, spawn_rayon_task, stopwatch_end, wait_or_graceful_shutdown, PerfStopwatch,
+};
 use futures::{stream, StreamExt};
 use num_traits::FromPrimitive;
 use starknet_providers::{ProviderError, SequencerGatewayProvider};
@@ -22,16 +31,6 @@ use starknet_types_core::felt::Felt;
 use tokio::sync::{mpsc, oneshot};
 use tokio::task::JoinSet;
 use tokio::time::Duration;
-
-use crate::commitments::compute_state_root;
-use crate::convert::{convert_and_verify_block, convert_and_verify_class};
-use crate::fetch::fetchers::{fetch_block_and_updates, FetchBlockId, L2BlockAndUpdates};
-use crate::fetch::l2_fetch_task;
-use crate::metrics::block_metrics::BlockMetrics;
-use crate::utility::trim_hash;
-use dp_utils::{
-    channel_wait_or_graceful_shutdown, spawn_rayon_task, stopwatch_end, wait_or_graceful_shutdown, PerfStopwatch,
-};
 
 // TODO: add more explicit error variants
 #[derive(thiserror::Error, Debug)]