+++ crates/client/sync/src/utils/utility.rs
@@ -1,15 +1,10 @@
 //! Utility functions for Deoxys.
 
-use anyhow::{bail, Context};
-use ethers::types::{I256, U256};
 use rand::seq::SliceRandom;
 use rand::thread_rng;
 use serde_json::Value;
-use starknet_api::hash::StarkFelt;
 use starknet_types_core::felt::Felt;
 use thiserror::Error;
-
-use crate::l1::{L1StateUpdate, LogStateUpdate};
 
 /// Returns a random Pokémon name.
 pub async fn get_random_pokemon_name() -> Result<String, Box<dyn std::error::Error>> {
@@ -22,25 +17,6 @@ pub async fn get_random_pokemon_name() -> Result<String, Box<dyn std::error::Err
     let random_pokemon = pokemon_array.choose(&mut rng).unwrap();
 
     Ok(random_pokemon["name"].as_str().unwrap().to_string())
-}
-
-pub fn u256_to_starkfelt(u256: U256) -> anyhow::Result<StarkFelt> {
-    let mut bytes = [0u8; 32];
-    u256.to_big_endian(&mut bytes);
-    StarkFelt::new(bytes).context("converting U256 to StarkFelt")
-}
-
-pub fn convert_log_state_update(log_state_update: LogStateUpdate) -> anyhow::Result<L1StateUpdate> {
-    let block_number = if log_state_update.block_number >= I256::zero() {
-        log_state_update.block_number.low_u64()
-    } else {
-        bail!("Block number is negative");
-    };
-
-    let global_root = u256_to_starkfelt(log_state_update.global_root)?;
-    let block_hash = u256_to_starkfelt(log_state_update.block_hash)?;
-
-    Ok(L1StateUpdate { block_number, global_root, block_hash })
 }
 
 #[derive(Error, Debug)]