+++ crates/node/src/service/sync.rs
@@ -1,27 +1,24 @@
-use std::sync::Arc;
-use std::time::Duration;
-
+use crate::cli::SyncParams;
+use alloy::primitives::Address;
 use anyhow::Context;
 use dc_db::db_metrics::DbMetrics;
 use dc_db::{DatabaseService, DeoxysBackend};
+use dc_eth::client::EthereumClient;
 use dc_metrics::MetricsRegistry;
 use dc_sync::fetch::fetchers::FetchConfig;
 use dc_sync::metrics::block_metrics::BlockMetrics;
 use dc_telemetry::TelemetryHandle;
-use primitive_types::H160;
 use starknet_types_core::felt::Felt;
+use std::sync::Arc;
+use std::time::Duration;
 use tokio::task::JoinSet;
-use url::Url;
-
-use crate::cli::SyncParams;
 
 #[derive(Clone)]
 pub struct SyncService {
     db_backend: Arc<DeoxysBackend>,
     fetch_config: FetchConfig,
     backup_every_n_blocks: Option<u64>,
-    l1_endpoint: Option<Url>,
-    l1_core_address: H160,
+    eth_client: EthereumClient,
     starting_block: Option<u64>,
     block_metrics: BlockMetrics,
     db_metrics: DbMetrics,
@@ -38,6 +35,7 @@ impl SyncService {
         metrics_handle: MetricsRegistry,
         telemetry: TelemetryHandle,
     ) -> anyhow::Result<Self> {
+        // TODO: create l1 metrics here
         let block_metrics = BlockMetrics::register(&metrics_handle)?;
         let db_metrics = DbMetrics::register(&metrics_handle)?;
         let fetch_config = config.block_fetch_config();
@@ -54,11 +52,15 @@ impl SyncService {
             None
         };
 
+        let core_address = Address::from_slice(config.network.l1_core_address().as_bytes());
+        let eth_client = EthereumClient::new(l1_endpoint.unwrap(), core_address, metrics_handle)
+            .await
+            .context("Creating ethereum client")?;
+
         Ok(Self {
             db_backend: Arc::clone(db.backend()),
             fetch_config,
-            l1_endpoint,
-            l1_core_address: config.network.l1_core_address(),
+            eth_client,
             starting_block: config.starting_block,
             backup_every_n_blocks: config.backup_every_n_blocks,
             block_metrics,
@@ -76,8 +78,7 @@ impl SyncService {
         let SyncService {
             fetch_config,
             backup_every_n_blocks,
-            l1_endpoint,
-            l1_core_address,
+            eth_client,
             starting_block,
             block_metrics,
             db_metrics,
@@ -88,12 +89,12 @@ impl SyncService {
         let telemetry = self.start_params.take().context("service already started")?;
 
         let db_backend = Arc::clone(&self.db_backend);
+
         join_set.spawn(async move {
             dc_sync::starknet_sync_worker::sync(
                 &db_backend,
                 fetch_config,
-                l1_endpoint,
-                l1_core_address,
+                eth_client,
                 starting_block,
                 backup_every_n_blocks,
                 block_metrics,