+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## Next release
 
+- refactor: L1BlockMetric is intialized inside the EthereumClient new function
+- refactor: BlockMetrics divided in L1BlockMetrics and BlockMetrics
+- test: tests added for eth client and event subscription
 - feat: add support for Starknet version 0.13.2
 - fix(l1): removed free l1 endpoint list
+- removed l1.rs from the sync crate and refactored it with alloy inside eth crate
+- refactor: removed l1.rs from the sync crate and refactored it with alloy inside eth crate
+- refactor: eth client (l1 crate) refactored with alloy
+- feat: added l1 crate (eth crate)
 - fix(metrics): removed influx and added l2_state_size data
 - fix: command to start the Madara client
 - refactor: database error unification